Move InstallationCandidate to models + add tests
